+++ app/src/main/AndroidManifest.xml
@@ -7,14 +7,14 @@
 
     <application
         android:allowBackup="true"
+        android:hardwareAccelerated="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme"
-        android:hardwareAccelerated="true" >
+        android:theme="@style/AppTheme">
         <activity
             android:name=".DashboardActivity"
-            android:configChanges="orientation|screenSize"/>
+            android:configChanges="orientation|screenSize" />
         <activity
             android:name=".MoviePlayerActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
@@ -54,10 +54,10 @@
         </activity>
         <activity
             android:name=".MovieDetailsActivity"
-            android:theme="@style/AppTheme"
+            android:configChanges="orientation|screenSize"
             android:label="@string/app_name"
-            android:configChanges="orientation|screenSize">
-        </activity>
+            android:theme="@style/AppTheme"></activity>
+        <activity android:name=".SearchActivity"></activity>
     </application>
 
 </manifest>
